Improve setup (WIP)

+++ WebinterfaceAPI/src/main/java/me/mrletsplay/webinterfaceapi/DefaultHandler.java
@@ -61,6 +61,7 @@ public class DefaultHandler implements ActionHandler {
 		String accountID = event.getData().getString("account");
 		Account acc = Webinterface.getAccountStorage().getAccountByID(accountID);
 		if(acc == null) return ActionResponse.error("Account doesn't exist");
+		if(acc.getID().equals(event.getAccount().getID())) return ActionResponse.error("Can't delete own account");
 		Webinterface.getAccountStorage().deleteAccount(acc);
 		return ActionResponse.success();
 	}
